Move withNewIMSI() into BaseTestCase.swift
